requirements.txt
gunicorn==19.9.0
absl-py==0.7.1
astor==0.8.0
Click==7.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
PyYAML==5.1
scipy==1.3.0
six==1.12.0
SQLAlchemy==1.3.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.1
termcolor==1.1.0
Werkzeug==0.15.4
